// src/generated/models/insurance-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InsuranceType } from './insurance-type';
// May contain unused imports in some cases
// @ts-ignore
import type { Relationship } from './relationship';

export interface InsuranceRequest {
    'hasInsurance': boolean;
    'insuranceType'?: InsuranceType;
    'schemeId'?: number;
    'policyCardNumber'?: string;
    'policyHolderName'?: string;
    'relationshipToHolder'?: Relationship;
    'policyExpiryDate'?: string;
    'insuranceCardFrontUrl'?: string;
    'insuranceCardBackUrl'?: string;
    'pmjayCardUrl'?: string;
    'claimAmountLimit'?: number;
}